Separate test example from modify example
