+++ examples/modify-test.html
@@ -9,7 +9,12 @@
     <link rel="stylesheet" href="../resources/bootstrap/css/bootstrap.min.css" type="text/css">
     <link rel="stylesheet" href="../resources/layout.css" type="text/css">
     <link rel="stylesheet" href="../resources/bootstrap/css/bootstrap-responsive.min.css" type="text/css">
-    <title>Modify features example</title>
+    <title>Modify features test</title>
+    <style>
+      .map {
+        background: grey;
+      }
+    </style>
   </head>
   <body>
 
@@ -32,11 +37,10 @@
       <div class="row-fluid">
 
         <div class="span12">
-          <h4 id="title">Modify features example</h4>
-          <p id="shortdesc">Editing features with the modify interaction.</p>
+          <h4 id="title">Modify features test</h4>
+          <p id="shortdesc">Example for testing feature modification.</p>
           <div id="docs">
-            <p>This example demonstrates how the modify and select interactions can be used together.  Zoom in to an area of interest and select a feature for editing.  Then drag points around to modify the feature.  You can preserve topology by selecting multiple features before editing (<code>Shift</code>+click to select multiple features).</p>
-            <p>See the <a href="modify-features.js" target="_blank">modify-features.js source</a> to see how this is done.</p>
+            <p>See the <a href="modify-test.js" target="_blank">modify-test.js source</a> to see how this is done.</p>
           </div>
           <div id="tags">modify, edit, vector</div>
         </div>
@@ -46,7 +50,7 @@
     </div>
 
     <script src="jquery.min.js" type="text/javascript"></script>
-    <script src="loader.js?id=modify-features" type="text/javascript"></script>
+    <script src="loader.js?id=modify-test" type="text/javascript"></script>
     <script src="../resources/example-behaviour.js" type="text/javascript"></script>
 
   </body>
